source/requirements.txt
openai==0.28.1
paho-mqtt==1.6.1